code refactor.
