+++ models/cls/model_manager.py
@@ -8,9 +8,12 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import torch
+
 from models.cls.nets.mobilenet import MobileNet
 from models.cls.nets.vgg import VGG
 from models.cls.nets.shufflenetv2 import ShuffleNetV2, ShuffleResNetV2
+from models.cls.loss.cls_modules import FCCELoss, FCCenterLoss
 from utils.tools.logger import Logger as Log
 
 CLS_MODEL_DICT = {
@@ -22,6 +25,11 @@ CLS_MODEL_DICT = {
     'shufflenetv2': ShuffleNetV2,
     'shufflenetv2-50': ShuffleResNetV2,
     'shufflenetv2-164': ShuffleResNetV2
+}
+
+CLS_LOSS_DICT = {
+    'fc_ce_loss': FCCELoss,
+    'fc_center_loss': FCCenterLoss
 }
 
 
@@ -40,3 +48,16 @@ class ModelManager(object):
         model = CLS_MODEL_DICT[model_name](self.configer)
 
         return model
+
+    def get_cls_loss(self, loss_type=None):
+        key = self.configer.get('loss', 'loss_type') if loss_type is None else loss_type
+        if key not in CLS_LOSS_DICT:
+            Log.error('Loss: {} not valid!'.format(key))
+            exit(1)
+
+        loss = CLS_LOSS_DICT[key](self.configer)
+        if self.configer.get('network', 'loss_balance') and len(range(torch.cuda.device_count())) > 1:
+            from extensions.tools.parallel.data_parallel import DataParallelCriterion
+            loss = DataParallelCriterion(loss)
+
+        return loss
