+++ models/det/model_manager.py
@@ -8,11 +8,14 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import torch
+
 from models.det.nets.darknet_yolov2 import DarkNetYolov2
 from models.det.nets.darknet_yolov3 import DarkNetYolov3
 from models.det.nets.vgg300_ssd import Vgg300SSD
 from models.det.nets.vgg512_ssd import Vgg512SSD
 from models.det.nets.faster_rcnn import FasterRCNN
+from models.det.loss.det_modules import SSDMultiBoxLoss, SSDFocalLoss, YOLOv3Loss, FRLoss
 from utils.tools.logger import Logger as Log
 
 DET_MODEL_DICT = {
@@ -21,6 +24,13 @@ DET_MODEL_DICT = {
     'darknet_yolov2': DarkNetYolov2,
     'darknet_yolov3': DarkNetYolov3,
     'faster_rcnn': FasterRCNN,
+}
+
+DET_LOSS_DICT = {
+    'ssd_multibox_loss': SSDMultiBoxLoss,
+    'ssd_focal_loss': SSDFocalLoss,
+    'yolov3_det_loss': YOLOv3Loss,
+    'fr_loss': FRLoss
 }
 
 
@@ -39,3 +49,16 @@ class ModelManager(object):
         model = DET_MODEL_DICT[model_name](self.configer)
 
         return model
+
+    def get_det_loss(self, loss_type=None):
+        key = self.configer.get('loss', 'loss_type') if loss_type is None else loss_type
+        if key not in DET_LOSS_DICT:
+            Log.error('Loss: {} not valid!'.format(key))
+            exit(1)
+
+        loss = DET_LOSS_DICT[key](self.configer)
+        if self.configer.get('network', 'loss_balance') and len(range(torch.cuda.device_count())) > 1:
+            from extensions.tools.parallel.data_parallel import DataParallelCriterion
+            loss = DataParallelCriterion(loss)
+
+        return loss
