+++ models/pose/model_manager.py
@@ -8,9 +8,12 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import torch
+
 from models.pose.nets.open_pose_org import get_open_pose_org
 from models.pose.nets.cpm_net import CPMNet
 from models.pose.nets.open_pose import OpenPose
+from models.pose.loss.pose_modules import OPMseLoss
 from utils.tools.logger import Logger as Log
 
 MULTI_POSE_MODEL_DICT = {
@@ -22,6 +25,9 @@ SINGLE_POSE_MODEL_DICT = {
     'cpm_net': CPMNet
 }
 
+POSE_LOSS_DICT = {
+    'op_mse_loss': OPMseLoss,
+}
 
 class ModelManager(object):
     def __init__(self, configer):
@@ -48,3 +54,16 @@ class ModelManager(object):
         model = SINGLE_POSE_MODEL_DICT[model_name](self.configer)
 
         return model
+
+    def get_pose_loss(self, loss_type=None):
+        key = self.configer.get('loss', 'loss_type') if loss_type is None else loss_type
+        if key not in POSE_LOSS_DICT:
+            Log.error('Loss: {} not valid!'.format(key))
+            exit(1)
+
+        loss = POSE_LOSS_DICT[key](self.configer)
+        if self.configer.get('network', 'loss_balance') and len(range(torch.cuda.device_count())) > 1:
+            from extensions.tools.parallel.data_parallel import DataParallelCriterion
+            loss = DataParallelCriterion(loss)
+
+        return loss
